Fix some TPD issues
1. Introduced Asserts for ZHEAP_METAPAGE before calling ZheapInitPage.
Metapage should not be initialized using this method.
2. Introduced Asserts to identify zheap metapage after reading the same.
3. Fixed tpd_desc.
4. In TPDPageAddEntry, we shouldn't shuffle the item pointers.
5. In TPDAllocatePageAndAddEntry, when new page is not added, we
shouldn't over-write previous and next block number in tpd opaque space.
6. Set LSN in page header for meta page.
7. For PageSetUNDO while locking a tuple, send set_tpd_map_slot=false.
8. Properly initilize fist and last tpd page in meta page.
9. In zheap_xlog_update, send proper undo pointer in TPDSetUNDO.
10. Fixed the storage of old tpd slot in xlrec.
11. Fixed compactify_ztuple memory overrun issue.

Patch by Dilip Kumar and me. Reviewed by Amit Kapila.

+++ src/include/access/zheapam_xlog.h
@@ -135,7 +135,7 @@ typedef struct xl_zheap_delete
 	/* zheap related info */
 	OffsetNumber offnum;		/* deleted tuple's offset */
 	uint16		infomask;	/* lock mode */
-	uint8		trans_slot_id;	/* transaction slot id */
+	uint16		trans_slot_id;	/* transaction slot id */
 	uint8		flags;
 } xl_zheap_delete;
 
@@ -178,7 +178,7 @@ typedef struct xl_zheap_update
 	/* zheap related info */
 	OffsetNumber old_offnum;	/* old tuple's offset */
 	uint16		old_infomask;	/* infomask bits to set on old tuple */
-	uint8		old_trans_slot_id;	/* old tuple's transaction slot id */
+	uint16		old_trans_slot_id;	/* old tuple's transaction slot id */
 	uint16		flags;
 	OffsetNumber new_offnum;	/* new tuple's offset */
 } xl_zheap_update;
@@ -223,7 +223,7 @@ typedef struct xl_zheap_lock
 	/* zheap related info */
 	OffsetNumber    offnum;		/* locked tuple's offset */
 	uint16	infomask;	/* lock mode */
-	uint8   trans_slot_id;		/* transaction slot id */
+	uint16   trans_slot_id;		/* transaction slot id */
 	uint8	flags;
 } xl_zheap_lock;
